+++ CHANGELOG.md
@@ -129,6 +129,7 @@ To create graph extensions, you now _must_ do so via a parent graph (using one o
 - **Fix**: Handle location-less declarations when reporting invalid assisted inject bindings.
 - **Fix**: Don't chunk parent graph validation statements as these must be in the original constructor body.
 - **Fix**: Fix wrong receiver context for chunked field initializers.
+- **Fix**: Fix support for generic private injected constructors.
 - [internal change] Simplify metadata and just use accessor annotations.
 - [internal change] Graph extension impls are now generated as nested classes within the generated metro graph that they are contributed to.